+++ headers/HashList.hpp
@@ -1,8 +1,28 @@
+/*
+ * HashList.hpp
+ *
+ * Hash Table
+ * A data structure template in C++
+ * Copyright (C) 2019 Chris Mabon
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 /**
- * @title Hash Table
- * @project hash_table
- * @github https://github.com/chrismabon/hash_table
+ * @package hash_table
  * @author Chris Mabon
+ * https://github.com/chrismabon
  */
 
 #ifndef HASH_TABLE_HASHLIST_HPP
@@ -12,38 +32,67 @@
 #include "KeyNode.hpp"
 
 /**
- * @brief Class HashList
+ * @class HashList
+ * @brief Linked list class to hold unique keys with common hashed value
  *
- * This is the linked list class to hold unique key nodes of the same
- * hashed value (hash collision).
+ * The hashed value is *NOT* stored in the list, it is computed as needed.
+ * head_key holds the first KeyNode appended to the list (unless the first node is removed
+ *      while other nodes are in the list, the second node will become the head node).
+ * Instantiation of HashList classes normally occurs only when a key is hashed and added to the table.
+ * Empty lists are the result of key deletion and are retained in the event they key is re-added.
  *
+ * OPERATIONS
+ *  -append key
+ *  -delete key
+ *  -find key
+ *  -clear list (delete hash)
  */
 class HashList {
-protected:
-    // Total collision entries for associated hash value
-    USI_t num_keys;
+private:
+    /// #####################
+    /// ### CLASS MEMBERS ###
+    /// #####################
 
-    // Unidirectional list of collision entries
-    KeyNode* keys;
+    /**
+     * LIST LENGTH
+     *  -Length of >1 indicates hash collision(s)
+     */
+    SmInt _numKeys;
+
+    /**
+     * HEAD NODE
+     *  -Can contain value nullptr indicating a deleted key
+     */
+    KeyNode* _headKey;
 
 public:
+    /// #################################
+    /// ### CONSTRUCTORS & DESTRUCTOR ###
+    /// #################################
+
+    // TODO finish comments on HashList header file
+    HashList();
+
     // Basic constructor
-    explicit HashList(USI_t num_keys = 0, KeyNode* keys = nullptr);
+    explicit HashList(SmInt numKeys = 0, KeyNode* keys = nullptr);
 
     // Destructor
     // Traverses the internal list of keys, deletes all entries
     virtual ~HashList();
 
     // Gets
-    USI_t get_num_keys() const;
+    SmInt getNumKeys() const;
 
-    KeyNode* get_keys() const;
+    KeyNode* getKeys() const;
 
     // Sets
-    void set_num_keys(USI_t i_num_keys);
+    void setNumKeys(SmInt numKeys);
 
-    void set_keys(KeyNode* i_keys);
+    void setHeadKey(KeyNode* keynode);
 
+    // List operations
+    // Add node
+    void appendNode(KeyNode* keynode);
 };