+++ source/Table.cpp
@@ -1,26 +1,46 @@
+/*
+ * Table.cpp
+ *
+ * Hash Table
+ * A data structure template in C++
+ * Copyright (C) 2019 Chris Mabon
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 /**
- * @title Hash Table
- * @project hash_table
- * @github https://github.com/chrismabon/hash_table
+ * @package hash_table
  * @author Chris Mabon
+ * https://github.com/chrismabon
  */
 
 #include "../headers/Table.hpp"
 
-Table::Table(ULLI_t i_fill, ULLI_t i_tbl_size, HashList** i_table)
+Table::Table(LgInt i_fill, LgInt i_tbl_size, HashList** i_table)
         : fill(i_fill), tbl_size(i_tbl_size), table(i_table) {
 
 }
 
-Table::~Table() {
+Table::~Table() = default {
 
 }
 
-ULLI_t Table::get_fill() const {
+LgInt Table::get_fill() const {
     return fill;
 }
 
-ULLI_t Table::get_tbl_size() const {
+LgInt Table::get_tbl_size() const {
     return tbl_size;
 }
 
@@ -28,11 +48,11 @@ HashList** Table::get_table() const {
     return table;
 }
 
-void Table::set_fill(ULLI_t i_fill) {
+void Table::set_fill(LgInt i_fill) {
     Table::fill = i_fill;
 }
 
-void Table::set_tbl_size(ULLI_t i_tbl_size) {
+void Table::set_tbl_size(LgInt i_tbl_size) {
     Table::tbl_size = i_tbl_size;
 }